Support Lumen (#7)
* Support Lumen

* Write README.md

* Tweak
