+++ composer.json
@@ -1,7 +1,7 @@
 {
     "name": "kra8/laravel-snowflake",
-    "description": "Snowflake for Laravel.",
-    "keywords": ["snowflake", "laravel"],
+    "description": "Snowflake for Laravel and Lumen.",
+    "keywords": ["snowflake", "Laravel", "Lumen"],
     "authors": [{
         "name": "Koki Asai",
         "email": "koki@asai.dev"
@@ -29,7 +29,7 @@
     "extra": {
         "laravel": {
             "providers": [
-                "Kra8\\Snowflake\\SnowflakeServiceProvider"
+                "Kra8\\Snowflake\\Providers\\LaravelServiceProvider"
             ]
         }
     }
