+++ src/Providers/LumenServiceProvider.php
@@ -1,0 +1,20 @@
+<?php
+
+namespace Kra8\Snowflake\Providers;
+
+class LumenServiceProvider extends AbstractServiceProvider
+{
+    /**
+     * Bootstrap any application services for lumen.
+     *
+     * @return void
+     */
+    public function boot()
+    {
+        $this->app->withFacades();
+
+        $path = realpath(__DIR__.'/../config/snowflake.php');
+
+        $this->mergeConfigFrom($path, 'snowflake');
+    }
+}
